+++ test/triggers/document-generated.test.js
@@ -15,12 +15,12 @@ const processesDocumentSample = {
 describe('Triggers::DocumentGenerated', () => {
   zapier.tools.env.inject();
 
-  describe('when only the App is specified', () => {
+  describe('when only the Workspace is specified', () => {
     it('returns a list of Documents', (done) => {
       pdfmonkeyApi
         .get('/api/v1/documents')
         .query({
-          'q[app_id]': '11111111-2222-3333-4444-555555555555',
+          'q[workspace_id]': '11111111-2222-3333-4444-555555555555',
           'page[size]': 100
         })
         .reply(200, { documents: [documentSample] });
@@ -28,7 +28,7 @@ describe('Triggers::DocumentGenerated', () => {
       const bundle = {
         ...bundleWithAuth(),
         inputData: {
-          appId: '11111111-2222-3333-4444-555555555555'
+          workspaceId: '11111111-2222-3333-4444-555555555555'
         }
       };
 
@@ -41,12 +41,12 @@ describe('Triggers::DocumentGenerated', () => {
     });
   });
 
-  describe('when both the App and Template are specified', () => {
+  describe('when both the Workspace and Template are specified', () => {
     it('returns a list of Documents', (done) => {
       pdfmonkeyApi
         .get('/api/v1/documents')
         .query({
-          'q[app_id]': '11111111-2222-3333-4444-555555555555',
+          'q[workspace_id]': '11111111-2222-3333-4444-555555555555',
           'q[document_template_id]': '22222222-3333-4444-5555-666666666666,33333333-4444-5555-6666-777777777777',
           'page[size]': 100
         })
@@ -55,7 +55,7 @@ describe('Triggers::DocumentGenerated', () => {
       const bundle = {
         ...bundleWithAuth(),
         inputData: {
-          appId: '11111111-2222-3333-4444-555555555555',
+          workspaceId: '11111111-2222-3333-4444-555555555555',
           documentTemplateId: [
             '22222222-3333-4444-5555-666666666666',
             '33333333-4444-5555-6666-777777777777'