+++ triggers/get-all-templates.js
@@ -8,7 +8,7 @@ const getAllTemplates = (z, bundle) => {
       Authorization: `Bearer ${bundle.authData.secretKey}`
     },
     params: {
-      'q[app_id]': bundle.inputData.appId
+      'q[workspace_id]': bundle.inputData.workspaceId
     }
   };
 
@@ -34,11 +34,11 @@ module.exports = {
     perform: getAllTemplates,
     inputFields: [
       {
-        key: 'appId',
-        label: 'App',
+        key: 'workspaceId',
+        label: 'Workspace',
         type: 'string',
         required: true,
-        dynamic: 'getAllApps.id.identifier'
+        dynamic: 'getAllWorkspaces.id.identifier'
       }
     ],
     sample: {